fix: render Asciidoc to `documentation`
This renders the Asciidoc content to the `documentation/manual`, which
is picked up by Hugo as static content.

The change needed to be made in the Gulp gulpfile.js in order to start
Hugo after the Asciidoc content is rendered to `documentation/manual` so
that Hugo is able to render it. Starting Hugo before the required static
directories exist renders the 404 even if the files are placed in the
directory after starting Hugo. This is because Hugo needs to setup file
watchers, and it cannot do that if the directory/files do not exist at
the start of `hugo serve`.